requirements.txt
pygame==2.1.2
Pillow==9.2.0
tkinter==0.1.0
